Emit binary directly with desired name instead of copying.
I'm getting test failures on Windows for

https://github.com/bazelbuild/rules_rust/pull/645

where the linker is complaining that it can't write to the output file.
The only reason I can imagine for this is that the test environment is
forbidding writes to anything other than the declared output files.

As I can't test on Windows directly myself, I regard this as an
experimental change that I'm pushing to a draft PR solely for the
purpose of checking whether this fixes the failing test.