Don't apply `name_to_crate_name` to test binary names.
Normal `rust_binary` targets don't do this, so for consistency, don't do
it in tests either.

An additional motivation for this change is that I'm trying to whittle down
the number of uses of the `name_to_crate_name` function, as it doesn't
take a possible `crate_name` attribute into account. For context, see
also the changes and discussion in

https://github.com/bazelbuild/rules_rust/pull/645